third_party/WebKit/LayoutTests/svg/animations/additive-from-to-fill-animation-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
PASS getComputedStyle(document.querySelector('rect')).fill is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE
